// FrontendQuiz/src/app/api/model/propertyType.ts
/**
 * SpeedQuiz API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const PropertyType = {
    Question: 'question',
    Answer: 'answer',
    Hint: 'hint',
    Image: 'image',
    Svg: 'svg',
    Audio: 'audio',
    Unknown: 'unknown'
} as const;
export type PropertyType = typeof PropertyType[keyof typeof PropertyType];