// FrontendQuiz/src/app/api/model/submissionDeck.ts
/**
 * SpeedQuiz API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PropertyType } from './propertyType';


export interface SubmissionDeck { 
    deckName: string;
    username: string;
    properties: { [key: string]: PropertyType; };
    cards: Array<{ [key: string]: string; }>;
}